chore: Add Dockerfile for running code validation
The code changes include the addition of a Dockerfile in the `docker/validator` directory. This Dockerfile sets up a Python environment, installs the required dependencies, copies the source code, and sets the working directory. It also adds an entrypoint to run the `validator.py` script. This Dockerfile is intended to be used for running code validation.